Added crit and arm pen formulae
Tweaked some stat mods and updated arm pen to be an integer.

ArmPen will work by negating x armour points from the enemy's armour
before calculating damage reduction from armour

+++ include/stats.h
@@ -56,7 +56,7 @@ struct Strength
     int base_strength;
     int strength;
     int attack_power;
-    float armour_pen;
+    int armour_pen;
     float parry_chance;
     float scale;
 };
